+++ blueprints/salle_bloc/views.py
@@ -2,12 +2,14 @@ from flask import Blueprint, request, jsonify
 
 from service.bloc_service import BlocService
 from service.salle_service import SalleService
+from tools.sql_injection_tools import SQLInjectionTools
 from tools.user_tools import UserTools
 
 
 class SalleBlocViews:
     def __init__(self):
         self.user_tools = UserTools()
+        self.sql_injection_tools = SQLInjectionTools()
         self.bloc_service = BlocService()
         self.salle_service = SalleService()
         self.salle_bloc_bp = Blueprint('salle_bloc', __name__, template_folder='templates')
@@ -19,6 +21,8 @@ class SalleBlocViews:
             if self.user_tools.check_user_in_session('admin'):
                 data = request.get_json()
                 bloc = data.get('nom_bloc')
+                if self.sql_injection_tools.detect_sql_injection([bloc]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
                 status = self.bloc_service.add_bloc(bloc)
                 if status != 'failed':
                     return {'status': 'success'}
@@ -30,6 +34,8 @@ class SalleBlocViews:
                 data = request.get_json()
                 bloc_id = data.get('id_bloc')
                 bloc_nom = data.get('nom_bloc')
+                if self.sql_injection_tools.detect_sql_injection([bloc_id, bloc_nom]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
                 status = self.bloc_service.update_bloc(bloc_id, bloc_nom)
                 if status != 'failed':
                     return {'status': 'success'}
@@ -40,6 +46,8 @@ class SalleBlocViews:
             if self.user_tools.check_user_in_session('admin'):
                 data = request.get_json()
                 bloc_id = data.get('id_bloc')
+                if self.sql_injection_tools.detect_sql_injection([bloc_id]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
                 status = self.bloc_service.delete_bloc(bloc_id)
                 if status != 'failed':
                     return {'status': 'success'}
@@ -58,6 +66,8 @@ class SalleBlocViews:
                 data = request.get_json()
                 salle = data.get('nom_salle')
                 bloc = data.get('id_bloc')
+                if self.sql_injection_tools.detect_sql_injection([salle, bloc]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
                 status = self.salle_service.add_salle(salle, bloc)
                 if status != 'failed':
                     return {'status': 'success'}
@@ -70,6 +80,8 @@ class SalleBlocViews:
                 salle_id = data.get('id_salle')
                 salle_nom = data.get('nom_salle')
                 bloc = data.get('id_bloc')
+                if self.sql_injection_tools.detect_sql_injection([salle_id, salle_nom, bloc]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
                 status = self.salle_service.update_salle(salle_id, salle_nom, bloc)
                 if status != 'failed':
                     return {'status': 'success'}
@@ -80,6 +92,8 @@ class SalleBlocViews:
             if self.user_tools.check_user_in_session('admin'):
                 data = request.get_json()
                 salle_id = data.get('id_salle')
+                if self.sql_injection_tools.detect_sql_injection([salle_id]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
                 status = self.salle_service.delete_salle(salle_id)
                 if status != 'failed':
                     return {'status': 'success'}
@@ -97,6 +111,8 @@ class SalleBlocViews:
             if self.user_tools.check_user_in_session('admin'):
                 data = request.get_json()
                 bloc = data.get('id_bloc')
+                if self.sql_injection_tools.detect_sql_injection([bloc]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
                 salles = self.salle_service.find_salle_by_bloc(bloc)
                 return {'status': 'success', 'salles': salles}
             return {'status': 'failed'}
